PR review
- euinquac model test file
- commented code left overs
- reference for speciesName
